agent/requirements.txt
phidata
openai
python-dotenv
